resolve build warnings in base layer for linux
resolve build warnings in base layer for linux
(unused parameter, comparison warning..)

Reviewed-on: https://gerrit.iotivity.org/gerrit/10895
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ resource/csdk/stack/src/oickeepalive.c
@@ -533,7 +533,7 @@ void IncreaseInterval(KeepAliveEntry_t *entry)
     VERIFY_NON_NULL_NR(entry, FATAL);
 
     OIC_LOG_V(DEBUG, TAG, "Total interval counts: %d", entry->intervalSize);
-    if (entry->intervalSize > entry->currIndex + 1)
+    if (entry->intervalSize > (size_t)entry->currIndex + 1)
     {
         entry->currIndex++;
         entry->interval = entry->intervalInfo[entry->currIndex];
